This time I actually did it :)
fixed enemy movement.

+++ Block.pde
@@ -1,11 +1,11 @@
 class Block {
   float x, y;
   PVector blockPos;
-  int blockSize;
+  float blockSize;
   int lastMove, speed;
   boolean release = false;
   
-  Block(PVector _blockPos, int _blockSize) {
+  Block(PVector _blockPos, float _blockSize) {
     this.blockPos = _blockPos;
     this.blockSize = _blockSize;
   }
